rename i10_pgm and correct some docstring
